requirements.txt
tensorflow==2.20.0
opencv-python==4.12.0.88
numpy>=2.1.0,<2.2.0
pandas==2.3.2
scikit-learn==1.7.1
matplotlib==3.10.5
sympy==1.14.0
threadpoolctl==3.6.0
pillow==11.3.0
keras==3.11.3
h5py==3.14.0
scipy==1.16.1
tqdm==4.67.1